Fixed some issues with FlowStatsReply. Added more length specs for protocol msgs.

+++ src/ofp/transmogrify.cpp
@@ -451,7 +451,7 @@ void Transmogrify::normalizeMPFlowReplyV1(size_t *start) {
   // Need at least 48 more bytes (136 - 88). If there's an action list, we need 
   // 8 more bytes for the instruction header.
   UInt16 actLen = UInt16_narrow_cast(length - 88);
-  int needed = 48;
+  size_t needed = 48;
   if (actLen > 0) {
     needed += 8;
   }
